// FrontEnd/ChamberAppel/src/app/generatedapis/models/RequeteAppelTempDatatableResponse.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
import type { RequeteAppelTemp } from './RequeteAppelTemp';
export type RequeteAppelTempDatatableResponse = {
    data?: Array<RequeteAppelTemp> | null;
    recordTotal?: number;
    recordFiltred?: number;
};